backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
langchain
crewai
anthropic
celery
redis
pydantic
pydantic-settings
yfinance
python-multipart
python-dotenv
httpx
beautifulsoup4
pandas
scipy
python-jose[cryptography]
passlib[bcrypt]
newsapi-python
alpha-vantage
langchain-google-genai
